+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added: `routingFallback` option for `openCookieClickerPage`.
+
 ## 0.3.1 - 2022-07-23
 
 - Changed: Cookie Connoisseur now downloads and uses the game files from the live version 2.048